requirements.txt
Bottleneck
certifi
chardet==3.0.4
click==8.1.3
colorama==0.4.5
cycler
Flaskrk
fonttools==4.25.0
gunicorn==20.1.0
idna==2.10
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==0.15.1
jsonify==0.5
kiwisolver
MarkupSafe
matplotlib
mkl-fft==1.3.1
mkl-random
mkl-service==2.4.0
munkres==1.1.4
numexpr
numpy
packaging
pandas
Pillow==9.0.1
pyparsing
python-dateutil
pytz
requests==2.24.0
scikit-learn==1.0.2
scipy
seaborn
sip==4.19.13
six
sklearn==0.0
threadpoolctl==2.1.0
tornado
typing_extensions
urllib3==1.25.9
Werkzeug==2.1.2
wincertstore==0.2
zipp==3.8.0
